test: extract MockConnectionManager to shared helper
Move _FakeDe1 (→ public FakeDe1) and MockConnectionManager out of
scan_step_test.dart into test/helpers/mock_connection_manager.dart so
launcher tests can share them without duplication.
